fix(Core/Scripting): Use uint8 for EnviromentalDamage in hook signatures
EnviromentalDamage is an unscoped enum without a fixed underlying type,
so it cannot be forward-declared. Use uint8 (matching AC pattern for
enum params in hook headers) and static_cast at the call site.

+++ src/server/game/Entities/Player/Player.cpp
@@ -775,7 +775,7 @@ uint32 Player::EnvironmentalDamage(EnviromentalDamage type, uint32 damage)
     if (IsImmuneToEnvironmentalDamage())
         return 0;
 
-    sScriptMgr->OnPlayerEnvironmentalDamage(this, type, damage);
+    sScriptMgr->OnPlayerEnvironmentalDamage(this, static_cast<uint8>(type), damage);
     if (damage == 0)
         return 0;
